PN-16823 Asset Changes
Port from PN-16823:
- CalibratedGLAccountId: INT -> BIGINT on AssetInventory/AssetInventoryAudit,
  matching the C# AssetInventory.CalibratedGLAccountId (long?) mapping.
- USP_GetWOAssetInventoryList: join AssetInventory->Asset to
  DeprNonDeprTangibleAssets instead of AssetAttributeType.
- Drop DATA_COMPRESSION = PAGE from AssetInventory perf indexes.

+++ PAS_Database/dbo/Tables/Tables1/AssetInventoryAudit.sql
@@ -172,7 +172,7 @@
     [ReceivablesAmount]                 DECIMAL (18, 6) NULL,
     [DepreciationStartDate]             DATETIME        NULL,
     [CalibrationCertificateNumber]      NVARCHAR (500)  NULL,
-    [CalibratedGLAccountId]             INT             NULL,
+    [CalibratedGLAccountId]             BIGINT          NULL,
     [CalibratedGLAccountName]           VARCHAR (255)   NULL,
     [DeprNonDeprTangibleAssetsId]       BIGINT          NULL,
     CONSTRAINT [PK_AssetInventoryAudit] PRIMARY KEY CLUSTERED ([AssetInventoryAuditId] ASC),
